Created SettingsTableViewController; ProviderStoryboard: adding a cell to edit profile in the menu which modally presents AccountViewController, which now has Cancel and Save buttons and no longer signs out; Settings cell modally presents SettingsTableViewController, which is where signing out happens now; Extensions: created UITableView hideEmptyCells()
