+++ HireMe/SharedController/AccountViewController.swift
@@ -6,8 +6,7 @@
 //  Copyright © 2016 AJ Bronson. All rights reserved.
 //
 
-import FBSDKLoginKit
-import GoogleSignIn
+import UIKit
 
 class AccountViewController: UITableViewController {
 	
@@ -16,7 +15,7 @@ class AccountViewController: UITableViewController {
     private var name: String?
     private var email: String?
     
-	//MARK: View controller life cycle
+	// MARK: - View controller life cycle
 	
 	override func viewDidLoad() {
 		super.viewDidLoad()
@@ -32,88 +31,43 @@ class AccountViewController: UITableViewController {
 	}
 	
     
-	//MARK: UITableViewDataSource callbacks
+	// MARK: - UITableViewDataSource
     
     override func numberOfSections(in tableView: UITableView) -> Int {
-        return 2
+        return 1
     }
 	
 	override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        switch section {
-        case 0: return 5
-        default: return 1
-        }
+        return 5
 	}
 	
 	override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        switch indexPath.section {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "accountCell", for: indexPath)
+        
+        switch indexPath.row {
         case 0:
-            let cell = tableView.dequeueReusableCell(withIdentifier: "accountCell", for: indexPath)
-            
-            switch indexPath.row {
-            case 0:
-                cell.textLabel?.text = "Name: \(self.name ?? "")"
-            case 1:
-                cell.textLabel?.text = "Skills: Tech, dogs, yard"
-            case 2:
-                cell.textLabel?.text = "Number: 888-888-8888"
-            case 3:
-                cell.textLabel?.text = "Email: \(self.email ?? "")"
-            default:
-                cell.textLabel?.text = "Change Password"
-            }
-            
-            return cell
+            cell.textLabel?.text = "Name: \(self.name ?? "")"
+        case 1:
+            cell.textLabel?.text = "Skills: Tech, dogs, yard"
+        case 2:
+            cell.textLabel?.text = "Number: 888-888-8888"
+        case 3:
+            cell.textLabel?.text = "Email: \(self.email ?? "")"
         default:
-            return tableView.dequeueReusableCell(withIdentifier: "logOutCell", for: indexPath)
+            cell.textLabel?.text = "Change Password"
         }
+        
+        return cell
 	}
     
     
-    // MARK: - UITableViewDelegate callbacks
+    // MARK: - IBActions
     
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-        
-        let reuseId = tableView.cellForRow(at: indexPath)?.reuseIdentifier
-        
-        if reuseId == "logOutCell" {
-            let alertController = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
-            alertController.popoverPresentationController?.sourceView = self.view
-            alertController.popoverPresentationController?.sourceRect = self.view.bounds;
-            alertController.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: nil))
-            alertController.addAction(UIAlertAction(title: "Sign Out", style: .destructive, handler: { (action) in
-                print("Signing out...") // DEBUG
-                switch getSignInMethod() {
-                    case .Facebook:
-                        FBSDKLoginManager().logOut()
-                        print("Signed out from Facebook") // DEBUG
-                    case .Google:
-                        GIDSignIn.sharedInstance().signOut()
-                        print("Signed out from Google") // DEBUG
-                    case .ThisApp:
-                        print("Signed out from LimitedHire") // DEBUG
-                    case .NotSignedIn:
-                        print("Not signed in") // DEBUG
-                }
-                
-                print("AccountViewController's parent \(self.parent?.descr)") // DEBUG
-                print("AccountViewController's parent's parent \(self.parent?.parent?.descr)") // DEBUG
-                print("presentingViewController \(self.presentingViewController?.descr)") // DEBUG
-                
-//                if let presentingVC = self.presentingViewController as? UINavigationController {
-//                    presentingVC.printViewControllers()
-//                    presentingVC.popToRootViewController(animated: true)
-//                    
-//                }
-                
-                if let parent = self.parent as? UINavigationController {
-                    parent.printViewControllers() // DEBUG
-                    parent.popToRootViewController(animated: true)
-                }
-            }))
-            
-            self.present(alertController, animated: true, completion: nil)
-        }
+    @IBAction func cancelTapped(_ sender: UIBarButtonItem) {
+        self.dismiss(animated: true, completion: nil)
+    }
+    
+    @IBAction func saveTapped(_ sender: UIBarButtonItem) {
+        print("Save profile")
     }
 }
